Silence several compiler warnings from VC9 x64
git-svn-id: https://svn.code.sf.net/p/apprenticevideo/code/trunk@520 28900df4-563f-46a4-9db8-70c0268b8f86

+++ apprenticevideo/yaeAPI.cpp
@@ -42,7 +42,7 @@ namespace yae
   // TTime::TTime
   //
   TTime::TTime(double seconds):
-    time_(1000000.0 * seconds),
+    time_((int64)(1000000.0 * seconds)),
     base_(1000000)
   {}
